+++ tests/test_routing.py
@@ -1165,11 +1165,13 @@ class TestRouting(unittest.TestCase):
         watershed_layer = None
 
     def test_single_drain_point(self):
-        """PGP.routing: test pitfilling."""
+        """PGP.routing: test single_drain_point pitfill."""
         dem_array = numpy.zeros((11, 11), dtype=numpy.float32)
         dem_array[3:8, 3:8] = -1.0
         dem_array[0, 0] = -1.0
         dem_array[10, 10] = -1.0
+        dem_path = os.path.join(self.workspace_dir, 'dem.tif')
+        _array_to_raster(dem_array, None, dem_path)
 
         # outlet tuple at 0,0, just drain one edge
         expected_array_0_0 = numpy.copy(dem_array)
@@ -1184,8 +1186,6 @@ class TestRouting(unittest.TestCase):
         for output_tuple, expected_array in [
                 ((0, 0), expected_array_0_0),
                 ((5, 5), expected_array_5_5)]:
-            dem_path = os.path.join(self.workspace_dir, 'dem.tif')
-            _array_to_raster(dem_array, None, dem_path)
             fill_path = os.path.join(self.workspace_dir, 'filled.tif')
             pygeoprocessing.routing.fill_pits(
                 (dem_path, 1), fill_path,
@@ -1193,3 +1193,28 @@ class TestRouting(unittest.TestCase):
                 working_dir=self.workspace_dir)
             result_array = pygeoprocessing.raster_to_numpy_array(fill_path)
             numpy.testing.assert_almost_equal(result_array, expected_array)
+
+    def test_detect_lowest_sink_and_drain(self):
+        """PGP>routing: test detect_lowest_sink_and_drain."""
+        dem_array = numpy.zeros((11, 11), dtype=numpy.float32)
+        dem_array[3:8, 3:8] = -1.0
+        dem_array[0, 0] = -1.0
+        dem_array[10, 10] = -1.0
+
+        dem_path = os.path.join(self.workspace_dir, 'dem.tif')
+        _array_to_raster(dem_array, None, dem_path)
+
+        edge_pixel, edge_height, pit_pixel, pit_height = \
+            pygeoprocessing.routing.detect_lowest_sink_and_drain(
+                (dem_path, 1))
+
+        expected_edge_pixel = (0, 0)
+        expected_edge_height = -1
+        expected_pit_pixel = (3, 3)
+        expected_pit_height = -1
+
+        self.assertEqual(edge_pixel, expected_edge_pixel)
+        self.assertEqual(edge_height, expected_edge_height)
+        print(pit_pixel)
+        self.assertEqual(pit_pixel, expected_pit_pixel)
+        self.assertEqual(pit_height, expected_pit_height)