+++ HISTORY.rst
@@ -75,6 +75,11 @@ Unreleased Changes
   forces a DEM to only have one outlet at any point on the raster. The
   fill effect is that all pixels will drain to the raster coordinate at
   ``single_outlet_tuple``.
+* Added a ``detect_lowest_sink_and_drain`` function that finds the lowest
+  DEM pixel that drains to nodata/edge and the lowest DEM pixel that could
+  be a sink. The values that result from this call can be used to condition
+  a DEM that is known to have a single drain using the
+  ``single_outlet_tuple`` parameter in ``routing.fill_pits``.
 
 2.1.2 (2020-12-03)
 ------------------